pp_channel : bugfix : 时间间隔转换问题，应该是float不是int

+++ pp_channel/fd_channel.py
@@ -207,7 +207,7 @@ class pp_toubiao_channel_manager(pp_thread):
                 global pp_global_info
                 while True:
                         try:
-                                time_interval = int(pp_global_info.timeout_channel[1])
+                                time_interval = float(pp_global_info.timeout_channel[1])
                         except:
                                 time_interval = 2
                         sleep(getsleeptime(time_interval))
